% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_func.R
\name{print.distr.seq.fit}
\alias{print.distr.seq.fit}
\title{Printing Distributed Sequential Estimation Result}
\usage{
\method{print}{distr.seq.fit}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"distr.seq.fit"}}

\item{...}{optional arguments to print methods.}
}
\description{
Print a distributed sequential estimation result
}
