% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_func.R
\name{print.seq.fit}
\alias{print.seq.fit}
\title{Printing Sequential Estimation Result}
\usage{
\method{print}{seq.fit}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"seq.fit"}}

\item{...}{optional arguments to print methods.}
}
\description{
Print a sequential estimation result
}
