% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_func.R
\name{summary.distr.seq.fit}
\alias{summary.distr.seq.fit}
\alias{print.summary.distr.seq.fit}
\title{Distributed Sequential Estimation Result Object}
\usage{
\method{summary}{distr.seq.fit}(object, ...)

\method{print}{summary.distr.seq.fit}(x, ...)
}
\arguments{
\item{object}{an object of class \code{"distr.seq.fit"}, usually, a result of a call to \code{\link{distr_seq_ana}}.}

\item{...}{additional arguments affecting the summary produced.}

\item{x}{an object of class \code{"summary.distr.seq.fit"}}
}
\description{
An object of class \code{"distr.seq.fit"} returned by \code{\link{distr_seq_ana}}.
It summarizes federated results across multiple sequential procedures.
}
\seealso{
\code{\link{distr_seq_ana}}, \code{\link{summary}}
}
