% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_func.R
\name{summary.seq.fit}
\alias{summary.seq.fit}
\alias{print.summary.seq.fit}
\title{Sequential Estimation Result Object}
\usage{
\method{summary}{seq.fit}(object, ...)

\method{print}{summary.seq.fit}(x, ...)
}
\arguments{
\item{object}{an object of class \code{"seq.fit"}, usually, a result of a call to \code{\link{seq_ana}}.}

\item{...}{additional arguments affecting the summary produced.}

\item{x}{an object of class \code{"summary.seq.fit"}}
}
\description{
An object of class \code{"seq.fit"} returned by \code{\link{seq_ana}}.
It contains the results of a sequential estimation procedure, including coefficient paths, information matrix, stopping info, and AUC (if applicable).
}
\seealso{
\code{\link{seq_ana}}, \code{\link{summary}}
}
